refactor: more renaming, and re-structure per module
