+++ src/index.ts
@@ -2,6 +2,5 @@ export * from './context';
 export * from './hooks';
 export * from './modules';
 export * from './types';
-export * from './error';
+export * from './purchaseError';
 export * from './eventEmitter';
-export * from './methods';
