+++ src/internal/linkingError.ts
