+++ src/internal/platform.ts
